feat: la fórmula del descuento usa los datos del cliente, no los del ejemplo
Último tramo. `recibos_de_alta` trabajaba con los valores de la transcripción
—50 % durante 90 días— porque no había de dónde sacar los del cliente.
`promocion_a_parametros` traduce lo que devuelve el transporte a los argumentos
de la fórmula, y con eso el cálculo deja de suponer.

Tres conversiones, ninguna trivial:

* PORCENTAJE A FRACCIÓN EXACTA. Llega como 50.0 y la fórmula quiere
  Fraction(1, 2). Se pasa por Fraction(str(...)): un float convertido directo
  arrastraría su error binario a la única aritmética del proyecto que se
  mantiene exacta a propósito.
* MESES A DÍAS. El CRM mide la promoción en meses y el recibo en DÍAS, que es
  lo que produce el recibo mixto. Con los 90 días por defecto, una promoción de
  6 meses partiría el cuarto recibo cuando quedan tres meses enteros.
* DÍAS YA CONSUMIDOS. El cliente no empieza de cero. Sin restarlos se le
  prometerían días de descuento que ya disfrutó.

Lo que no consta no se inventa: un campo ausente se queda fuera y la fórmula
usa su valor por defecto. «No consta el porcentaje» y «el descuento es del 0 %»
son cosas distintas y solo una es cierta.

Comprobado con el registro real de la cuenta 759816134 —RA, 6 meses, 50 %, 31
días consumidos—: da 149 días de bolsa y una serie de cuatro recibos a mitad de
precio, uno partido y vuelta al precio de lista, con la bolsa a cero.

De paso, el importe del descuento se convierte en Decimal y no en float, que es
la regla del proyecto para todo lo que sea dinero.

+++ apps/api/transporte_supabase.py
@@ -45,6 +45,7 @@ import os
 from collections import defaultdict
 from collections.abc import Mapping
 from datetime import date, timedelta
+from decimal import Decimal
 from typing import Any
 
 from apps.api.acl import CuentaNoEncontradaExterna, ErrorSistemaExterno
@@ -492,7 +493,7 @@ class TransporteSupabase:
             "dias_consumidos": int(dias) if dias is not None else None,
             "cuota_numero": int(fila[5]) if fila[5] is not None else None,
             "monto_descuento_cent": (
-                int(round(float(fila[6]) * 100)) if fila[6] is not None else None
+                round(Decimal(fila[6]) * 100) if fila[6] is not None else None
             ),
             "descripcion": fila[7] or "",
             "tipo_descuento": fila[8] or "",
